requirements.txt
asposestorage==1.0.2
loguru==0.7.2
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
Pillow==9.5.0
Pillow==10.2.0
scikit_learn==1.0.2
scipy==1.7.3
six==1.16.0
torch==1.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.66.1
visualize==0.5.1